# PropertyRentalSpring/src/main/java/ru/yuriy/propertyrental/util/response_body/ErrorResponse.java
package ru.yuriy.propertyrental.util.response_body;

import com.fasterxml.jackson.annotation.JsonFormat;
import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

import java.sql.Timestamp;

@Getter
@Setter
@Builder
public class ErrorResponse
{
    private String status;

    private Integer code;

    private String errorMessage;

    @JsonFormat(shape = JsonFormat.Shape.STRING,
            pattern = "yyyy-MM-dd HH:mm:ss", timezone = "UTC")
    private Timestamp timestamp;
}